+++ cart/views.py
@@ -6,6 +6,7 @@ from rest_framework.filters import SearchFilter
 from rest_framework.response import Response
 from rest_framework.viewsets import ModelViewSet
 
+from cart.models import Cart
 from cart.models import CartItem
 from cart.paginators import CartItemPagination
 from cart.paginators import CartPagination
@@ -17,6 +18,7 @@ from core.filters.custom_filters import PascalSnakeCaseOrderingFilter
 
 
 class CartViewSet(ModelViewSet):
+    queryset = Cart.objects.all()
     serializer_class = CartSerializer
     pagination_class = CartPagination
     filter_backends = [DjangoFilterBackend, PascalSnakeCaseOrderingFilter, SearchFilter]
@@ -29,41 +31,43 @@ class CartViewSet(ModelViewSet):
         "user",
     ]
 
-    def get_queryset(self):
-        service = CartService(self.request)
-        return service.cart
-
     def get_serializer_context(self):
         context = super().get_serializer_context()
-        cart_service = CartService(self.request)
+        cart_service = CartService(request=self.request)
         context["cart"] = cart_service.cart
         return context
 
     def retrieve(self, request, *args, **kwargs) -> Response:
-        service = CartService(self.request)
-        serializer = self.get_serializer(service.cart)
+        service = CartService(request=request)
+        serializer = self.get_serializer(service.get_or_create_cart(request))
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def update(self, request, *args, **kwargs) -> Response:
-        service = CartService(self.request)
-        serializer = self.get_serializer(service.cart, data=request.data)
+        service = CartService(request=request)
+        serializer = self.get_serializer(
+            service.get_or_create_cart(request), data=request.data
+        )
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_200_OK)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def partial_update(self, request, *args, **kwargs) -> Response:
-        service = CartService(self.request)
-        serializer = self.get_serializer(service.cart, data=request.data, partial=True)
+        service = CartService(request=request)
+        serializer = self.get_serializer(
+            service.get_or_create_cart(request), data=request.data, partial=True
+        )
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_200_OK)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def destroy(self, request, *args, **kwargs) -> Response:
-        service = CartService(self.request)
-        service.cart.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        service = CartService(request=request)
+        if service.cart:
+            service.cart.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(status=status.HTTP_404_NOT_FOUND)
 
 
 class CartItemViewSet(ModelViewSet):
@@ -82,11 +86,15 @@ class CartItemViewSet(ModelViewSet):
     def get_queryset(self):
         if getattr(self, "swagger_fake_view", False):
             return CartItem.objects.none()
+
+        if self.request.user.is_anonymous:
+            return CartItem.objects.filter(cart__id=self.request.session.get("cart_id"))
         return CartItem.objects.filter(cart__user=self.request.user)
 
     def get_serializer_context(self):
         context = super().get_serializer_context()
-        context["cart"] = CartService(self.request).cart
+        service = CartService(request=self.request)
+        context["cart"] = service.get_or_create_cart(self.request)
         return context
 
     def list(self, request, *args, **kwargs):
@@ -96,7 +104,8 @@ class CartItemViewSet(ModelViewSet):
 
     def create(self, request, *args, **kwargs):
         serializer = CartItemCreateSerializer(data=request.data)
-        serializer.context["cart"] = CartService(self.request).cart
+        service = CartService(request=request)
+        serializer.context["cart"] = service.get_or_create_cart(request)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
